refactor: move network proxy to modules/network/ sub-package
Move all 30 network proxy files (13 production + 17 test) from modules/
root into modules/network/ sub-package. This creates a symmetric layout
where both proxies live under their own sub-packages:
- modules/network/ (network/RPC proxy)
- modules/ai/ (AI proxy)

The only external consumer (module.go) is updated to import the new path.
Caddy module IDs are unchanged since they're set by CaddyModule() methods.
