+++ modules/network/caddy_unmarshaller.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"encoding/hex"
