+++ modules/network/din_middleware.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"bytes"
