+++ modules/network/din_middleware_helpers.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"crypto/sha256"
