+++ modules/network/din_provider_filter_test.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"testing"
