+++ modules/network/din_scorebased_selector.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"net/http"
