+++ modules/network/din_scorebased_selector_test.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"context"
