+++ modules/network/din_select.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"net/http"
