+++ modules/network/din_select_test.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"context"
