+++ modules/network/din_upstreams.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"fmt"
