+++ modules/network/handler_set_test.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"testing"
