+++ modules/network/helpers.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"bytes"
