+++ modules/network/network.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"container/list"
