+++ modules/network/network_block_lag.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"math"
