+++ modules/network/network_block_lag_test.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"encoding/json"
