+++ modules/network/network_evaluate_health_test.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"testing"
