+++ modules/network/network_get_block_test.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"net/http"
