+++ modules/network/response_writer_wrapper.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"bytes"
