+++ modules/network/response_writer_wrapper_test.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"net/http/httptest"
