+++ modules/network/rest_middleware_flow_test.go
@@ -1,4 +1,4 @@
-package modules
+package network
 
 import (
 	"container/list"
